Generate SafeInvoke overloads with T4
